+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `aws.giantswarm.io/dns-mode` annotation.
+- Add `aws.giantswarm.io/dns-assign-additional-vpc` annotation.
+
 ## [0.16.1] - 2022-12-02
 
 ### Fixed